Food search endpoint queries edemam
Endpoint /api/v1/recipes/food_search?q=<liver> queries the database for any recipes with "liver" as the foodType. If it doesn't find any, it makes a fetch call to the Edemam API. It then checks that it got any hits. If it did, it inserts them into the database.

Co-authored-by: Tim Allen <tnelsallen3@gmail.com>